docker/betterme.news-docker/betterme/requirements.txt
app==0.0.1
beautifulsoup4==4.11.1
discord.py==1.7.3
fastapi==0.78.0
imgbbpy==0.1.3
numpy==1.21.5
opencv_python==4.5.5.64
pydantic==1.9.0
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.27.1
uvicorn==0.17.6
waiting==1.4.1
wget==3.2
Zenora==0.0.3